chore: provide helper for determining allowed roles
