+++ packages/client/src/v2-events/utils.ts
@@ -24,7 +24,8 @@ import {
   getAcceptedScopesByType,
   decodeScope,
   RecordScopeTypeV2,
-  EncodedScope
+  EncodedScope,
+  getAdministrativeAreaHierarchy
 } from '@opencrvs/commons/client'
 
 export function getUsersFullName(name: UserOrSystem['name']) {
@@ -190,33 +191,6 @@ export function mergeWithoutNullsOrUndefined<T>(
 }
 
 type OutputMode = 'withIds' | 'withNames'
-
-// Given an administrative area id, return the full hierarchy from root to leaf.
-export function getAdministrativeAreaHierarchy(
-  administrativeAreaId: string | undefined | null,
-  administrativeAreas: Map<UUID, AdministrativeArea>
-) {
-  // Collect location objects from leaf to root
-  const collectedLocations: AdministrativeArea[] = []
-
-  const parsedAdministrativeAreaId =
-    administrativeAreaId && UUID.safeParse(administrativeAreaId).data
-
-  let current = parsedAdministrativeAreaId
-    ? administrativeAreas.get(parsedAdministrativeAreaId)
-    : null
-
-  while (current) {
-    collectedLocations.push(current)
-    if (!current.parentId) {
-      break
-    }
-    const parentId = current.parentId
-    current = administrativeAreas.get(parentId)
-  }
-
-  return collectedLocations
-}
 
 /*
   Function to traverse the administrative level hierarchy from an arbitrary / leaf point
