+++ packages/commons/src/client.ts
@@ -22,6 +22,7 @@ export * from './authentication'
 export * from './url'
 export * from './events/serializers/user/serializer'
 export * from './events/serializers/date/serializer'
+export * from './events/mocks.test.utils'
 export { TriggerEvent } from './notification'
 export * from './application-config'
 export * from './icons'
