+++ packages/commons/src/events/locations.ts
@@ -324,3 +324,68 @@ export function userCanAccessEventWithScopes(
     canAccessEventWithScope(event, scope, user, customActionType)
   )
 }
+
+// Given an administrative area id, return the full hierarchy from root to leaf.
+export function getAdministrativeAreaHierarchy(
+  administrativeAreaId: string | undefined | null,
+  administrativeAreas: Map<UUID, AdministrativeArea>
+) {
+  // Collect location objects from leaf to root
+  const collectedLocations: AdministrativeArea[] = []
+
+  const parsedAdministrativeAreaId =
+    administrativeAreaId && UUID.safeParse(administrativeAreaId).data
+
+  let current = parsedAdministrativeAreaId
+    ? administrativeAreas.get(parsedAdministrativeAreaId)
+    : null
+
+  while (current) {
+    collectedLocations.push(current)
+    if (!current.parentId) {
+      break
+    }
+    const parentId = current.parentId
+    current = administrativeAreas.get(parentId)
+  }
+
+  return collectedLocations
+}
+
+/**
+ * Resolves a location or administrative area UUID into a root-first hierarchy of UUIDs.
+ *
+ * If the ID refers to a `Location` (e.g. CRVS office), the hierarchy includes
+ * the administrative area ancestors followed by the location itself.
+ * If the ID refers to an `AdministrativeArea`, returns the area's ancestor chain (root-first).
+ *
+ * Uses `getAdministrativeAreaHierarchy` which returns leaf-first order,
+ * so the result is reversed to match the root-first convention used by the server.
+ */
+export function getLocationHierarchy(
+  selectedId: UUID,
+  context: {
+    administrativeAreas: Map<UUID, AdministrativeArea>
+    locations: Map<UUID, Location>
+  }
+): UUID[] {
+  const { administrativeAreas, locations } = context
+  const loc = locations.get(selectedId)
+
+  if (loc) {
+    if (loc.administrativeAreaId) {
+      const hierarchy = getAdministrativeAreaHierarchy(
+        loc.administrativeAreaId,
+        administrativeAreas
+      )
+      return [...hierarchy.reverse().map((area) => area.id), loc.id]
+    }
+    return [loc.id]
+  }
+
+  const hierarchy = getAdministrativeAreaHierarchy(
+    selectedId,
+    administrativeAreas
+  )
+  return hierarchy.reverse().map((area) => area.id)
+}
